Fixing test to correspond changes

+++ master/buildbot/db/migrate/versions/018_add_sourcestampset.py
@@ -39,8 +39,10 @@ def upgrade(migrate_engine):
     migrate_engine.execute(str(sautils.InsertFromSelect(sourcestampsets_table, sourcestampsetids)))
 
     # rename the buildsets table column
-    cons = constraint.ForeignKeyConstraint([buildsets_table.c.sourcestampid], [sourcestamps_table.c.id])
-    cons.drop()
+    if migrate_engine.dialect.name != 'sqlite':
+        cons = constraint.ForeignKeyConstraint([buildsets_table.c.sourcestampid], [sourcestamps_table.c.id])
+        cons.drop()
+
     buildsets_table.c.sourcestampid.alter(name='sourcestampsetid')
 
     metadata.remove(buildsets_table)
